\name{breast.17}
\docType{data}
\alias{breast.17}
\title{Aspirin and Breast Cancer: 17 studies}
\description{This dataset gives log odds of
  breast cancer for long-term aspirin users, and its standard error,
  derived from 17 cohort and case-control studies.}
\usage{data(breast.17)}
\format{A data frame with seventeen rows, corresponding to the seventeen
  papers.  There are two columns: psi.hat (numeric, the log odds ratio),
  and se.psi.hat (numeric, estimated SE of the log odds ratio).
  The \code{rownames} attribute gives the first author of the paper
  and the citation number of the study in \cite{Harris et.\ al. (2005)}.}
\source{Harris, R., Beebe-Donk, J., Doss, H., and Burr, D. (2005).
  \dQuote{Aspirin, ibuprofen, and other non-steroidal anti-inflammatory
  drugs in cancer prevention: A critical review of non-selective
  COX-2 blockade (Review).}  \emph{Oncology Reports 13} 559-583.}
\references{
  Gonzalez-Perez, A., Rodriguez, L., and Lopez-Ridaura, R. (2003).
   \dQuote{Effects of non-steroidal anti-inflammatory drugs
    on cancer sites other than the colon and rectum: a meta-analysis.}
  \emph{BMC Cancer 3} 28.
  
  Khuder, S. and Mutgi, A. (2001).  \dQuote{Breast cancer and
   NSAID use: a meta-analysis.}  \emph{British Journal of Cancer 84},
   1188--1192.
  }

  
