\name{ddtm.s}
\docType{data}
\alias{ddtm.s}
\title{Decontamination of the Digestive Tract Mortality, Short Dataset}
\description{This dataset gives number of deaths and number of subjects
  in treatment vs. control groups in 14 studies from a meta-analysis of
  studies on antiobiotics to reduce infection in intensive-care units.}
\usage{data(ddtm.s)}
\format{A data frame with fourteen rows, corresponding to fourteen
  different, unidentified studies.  There are four columns: number of
  deaths in the treatment group, sample size in the treatment group,
  number of deaths in the control group, sample size in the control
  group.

  \cite{Burr and Doss (2005)} give the background for this dataset.}
\source{Selective Decontamination of the Digestive Tract Trialists'
  Collaborative Group (1993).  \dQuote{Meta-analysis of randomised
  controlled trials of selective decontamination of the digestive tract.}
  \emph{British Medical Journal 307} 525--532.}
\references{
Burr, D. and Doss, H. (2005).  \dQuote{A Bayesian semiparametric model
for random-effects meta-analysis.}  \emph{Journal of the American
Statistical Association 100} 242--251.  
}
