\name{print.dir.cond}
\alias{print.dir.cond}
\title{printing method for objects of class dir.cond}
\description{gives limited info}
\usage{
\method{print}{dir.cond}(x,digits=max(3,getOption("digits") -3),\dots)
}
\arguments{
  \item{x}{object of class dir.cond}
  \item{digits}{positive integer, number of digits for printing}
  \item{\dots}{place holder}
}  
